refactor(healthplatform): update collector and check runner to new component paths
Updates pkg/collector (runner, worker, check/stats) and comp/collector
to import from comp/healthplatform/core/def instead of the old layout.
