+++ pkg/collector/check/stats/stats_test.go
@@ -14,8 +14,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/DataDog/datadog-agent/comp/core/telemetry/impl"
-	healthplatformmock "github.com/DataDog/datadog-agent/comp/healthplatform/mock"
+	"github.com/DataDog/datadog-agent/comp/core/telemetry/telemetryimpl"
+	healthplatformmock "github.com/DataDog/datadog-agent/comp/healthplatform/core/mock"
 	checkid "github.com/DataDog/datadog-agent/pkg/collector/check/id"
 	configmock "github.com/DataDog/datadog-agent/pkg/config/mock"
 )
